refactor(renderer): isolate interactive auth behind injected native API [skip ci]
