+++ desktop/renderer/features/auth-challenge/index.mjs
@@ -1,0 +1,1 @@
+export { createAuthChallengeFeature, MAX_RESPONSE_BYTES, start } from './controller.mjs';
